Use buffer and write to binary output
Replaced easy-read matrix repesentation with unsigned char buffer to write to binary file,
though one issue is that previously in part 1 we were gauranteed that the values would be
less than 256. With the matrix decomposition, this might not be so. Dr. Duan briefly spoke
on this regarding calculating our potential error based on the size of our data storage.